requirements.txt
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
distlib==0.3.6
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
drf-base64==2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.12.0
h11==0.14.0
idna==3.4
jdcal==1.4.1
openpyexcel==2.5.14
openpyxl==3.1.2
outcome==1.2.0
Pillow==9.5.0
platformdirs==3.5.1
pycparser==2.21
PyJWT==2.7.0
pymongo==4.3.3
PySocks==1.7.1
pytz==2023.3
redis==4.5.5
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.2.4
trio==0.22.0
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.0
virtualenvwrapper-win==1.2.7
wsproto==1.2.0
